used module and made the js files small unite
