+++ index.html
@@ -17,6 +17,7 @@
       rel="stylesheet"
       href="https://unpkg.com/boxicons@latest/css/boxicons.min.css"
     />
+    <script type="module" src="/script/nav.js"></script>
   </head>
   
   <body>
@@ -205,6 +206,5 @@
         </div>
       </section>
     </div>
-    <script src="/script/nav.js"></script>
   </body>
 </html>
