+++ index/Payment.html
@@ -21,6 +21,8 @@
     <link rel="stylesheet" href="/src/css/navigation.css" />
     <link rel="stylesheet" href="/src/css/payment.css" />
     <link rel="stylesheet" href="/src/css/footer.css" />
+    <script type="module" src="/script/nav.js"></script>
+    <script src="/script/payment.js" defer></script>
   </head>
 
   <body>
@@ -157,7 +159,6 @@
         </div>
       </section>
     </div>
-    <script src="/script/nav.js"></script>
     <script
       type="module"
       src="https://unpkg.com/ionicons@7.1.0/dist/ionicons/ionicons.esm.js"
@@ -166,6 +167,5 @@
       nomodule
       src="https://unpkg.com/ionicons@7.1.0/dist/ionicons/ionicons.js"
     ></script>
-    <script src="/script/payment.js"></script>
   </body>
 </html>
