+++ index/aboutPage.html
@@ -21,6 +21,7 @@
     <link rel="stylesheet" href="/src/css/navigation.css" />
     <link rel="stylesheet" href="/src/css/aboutPage.css" />
     <link rel="stylesheet" href="/src/css/footer.css" />
+    <script type="module" src="/script/nav.js"></script>
   </head>
   <body>
     <div class="container">
@@ -154,6 +155,5 @@
         </div>
       </section>
     </div>
-    <script src="/script/nav.js"></script>
   </body>
 </html>
