+++ index/cartInfo.html
@@ -21,6 +21,8 @@
     <link rel="stylesheet" href="/src/css/navigation.css" />
     <link rel="stylesheet" href="/src/css/cartInfo.css" />
     <link rel="stylesheet" href="/src/css/footer.css" />
+    <script type="module" src="/script/nav.js"></script>
+    <script type="module" src="/script/cartInfo.js"></script>
   </head>
 
   <body>
@@ -176,7 +178,5 @@
         </div>
       </section>
     </div>
-    <script src="/script/nav.js"></script>
-    <script type="module" src="/script/cartInfo.js"></script>
   </body>
 </html>
