+++ index/contactPage.html
@@ -21,6 +21,7 @@
     <link rel="stylesheet" href="/src/css/navigation.css" />
     <link rel="stylesheet" href="/src/css/contactPage.css" />
     <link rel="stylesheet" href="/src/css/footer.css" />
+    <script type="module" src="/script/nav.js"></script>
   </head>
 
   <body>
@@ -159,6 +160,5 @@
         </div>
       </section>
     </div>
-    <script src="/script/nav.js"></script>
   </body>
 </html>
