+++ index/delivery.html
@@ -21,6 +21,8 @@
     <link rel="stylesheet" href="/src/css/navigation.css" />
     <link rel="stylesheet" href="/src/css/delivery.css" />
     <link rel="stylesheet" href="/src/css/footer.css" />
+    <script type="module" src="/script/nav.js"></script>
+    <script src="/script/delivery.js" defer></script>
   </head>
 
   <body>
@@ -81,8 +83,11 @@
                 </a>
               </p>
             </div>
-            <button class="submitButton" id="submitButton">Proceed to checkout</button>
+            <button class="submitButton" id="submitButton">
+              Proceed to checkout
+            </button>
           </div>
+        </div>
       </section>
     </div>
     <div class="container">
@@ -155,7 +160,5 @@
         </div>
       </section>
     </div>
-    <script src="/script/nav.js"></script>
-    <script src="/script/delivery.js"></script>
   </body>
 </html>
