+++ index/productPage.html
@@ -21,8 +21,10 @@
       rel="stylesheet"
       href="https://unpkg.com/boxicons@latest/css/boxicons.min.css"
     />
-    <script src="/script/nav.js" defer></script>
+    <script type="module" src="/script/nav.js"></script>
     <script type="module" src="/script/main.js"></script>
+    <script type="module" src="/script/fetchProductDetails.js"></script>
+    
   </head>
   <body>
     <div class="container">
