+++ index/productSpecificPage.html
@@ -25,6 +25,8 @@
       rel="stylesheet"
       href="https://unpkg.com/boxicons@latest/css/boxicons.min.css"
     />
+    <script type="module" src="../script/productSpecific.js"></script>
+    <script type="module" src="/script/nav.js"></script>
   </head>
   <body>
     <div class="container">
@@ -143,9 +145,5 @@
         </div>
       </section>
     </div>
-    <script src="/script/nav.js"></script>
-
-    <script type="module" src="../script/productSpecific.js"></script>
-    <script src="/script/productSpecificPage.js"></script>
   </body>
 </html>
