+++ index/successPage.html
@@ -21,6 +21,8 @@
     <link rel="stylesheet" href="/src/css/navigation.css" />
     <link rel="stylesheet" href="/src/css/successPage.css" />
     <link rel="stylesheet" href="/src/css/footer.css" />
+    <script type="module" src="/script/nav.js"></script>
+    <script src="/script/successPage.js" defer></script>
   </head>
 
   <body>
@@ -135,7 +137,5 @@
         </div>
       </section>
     </div>
-    <script src="/script/nav.js"></script>
-    <script src="/script/successPage.js"></script>
   </body>
 </html>
