--- script/app.js
@@ -1,4 +1,0 @@
-import { renderProducts } from '../script/products.js';
-
-// Call the renderProducts function
-renderProducts();
