+++ script/getClickedProductId.js
@@ -1,0 +1,8 @@
+// Function to retrieve the clicked product ID from the URL parameters
+function getClickedProductId() {
+  const queryString = window.location.search;
+  const urlParams = new URLSearchParams(queryString);
+  return urlParams.get("id");
+}
+
+export { getClickedProductId };
