+++ script/main.js
@@ -1,5 +1,5 @@
 import { loginUser } from "../script/user.js";
-import { renderGenderFilter, filterProducts } from "./filter.js";
+import { renderGenderFilter } from "./filter.js";
 import { renderFilteredProducts } from "./productRenderer.js";
 import { showLoader, hideLoader } from "./loader.js";
 
@@ -8,7 +8,6 @@ let clickedProduct = null;
 
 async function renderProducts() {
   try {
-    // Show loader while fetching products
     showLoader();
 
     const token = await loginUser(
@@ -34,11 +33,9 @@ async function renderProducts() {
     renderGenderFilter(products); // Call the gender filter rendering function here
     renderFilteredProducts(products); // Render all products initially
 
-    // Hide loader after products are rendered
     hideLoader();
   } catch (error) {
     console.error("Error rendering products:", error);
-    // Hide loader in case of error
     hideLoader();
   }
 }
